+++ autokeras/__init__.py
@@ -1,2 +1,2 @@
-from autokeras.image_supervised import ImageClassifier, ImageRegressor
+from autokeras.image.image_supervised import ImageClassifier, ImageRegressor
 from autokeras.text.text_supervised import TextClassifier
