+++ tests/image/test_image_supervised.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from autokeras.image_supervised import *
+from autokeras.image.image_supervised import *
 from tests.common import clean_dir, MockProcess, simple_transform
 
 
@@ -175,7 +175,7 @@ def test_fit_csv_file(_, _1):
     clean_dir(os.path.join(path, "temp"))
 
 
-@patch('autokeras.image_supervised.temp_folder_generator', return_value='dummy_path/')
+@patch('autokeras.image.image_supervised.temp_folder_generator', return_value='dummy_path/')
 def test_init_image_classifier_with_none_path(_):
     clf = ImageClassifier()
     assert clf.path == 'dummy_path/'
