+++ tests/text/test_text_preprocessor.py
